Added Singleton Global type to additional function control
Adding a singleton Global type in order to manipualte accorss the application some functionalities that can be propagated to execution threads. One sample is added in the Test.fs file on how this might be implemented.

Another addition in here is a test for the output led.

At this point I did some changes in the physical circuit since I clean up a bit the wiring and I had some issues because of bad connections or a pending reboot on the PI, but it got resolved and is working up to this point.